requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
requests==2.32.3
python-dotenv==1.0.1
scikit-learn==1.5.2
numpy==2.1.1
pypdf>=4
